feat(avo-2597): Sort users before looping, refactor
