--- src/shared/components/ShareDropdown/ShareDropdown.helpers.tsx
@@ -1,5 +1,0 @@
-import { ShareUserInfoRights } from './ShareDropdown.types';
-
-export const shareUserRightToString = (right: ShareUserInfoRights) => {
-	return (right as unknown as string).toLocaleLowerCase();
-};
