+++ src/shared/components/ShareWithColleagues/ShareWithColleagues.types.ts
